+++ javascript/node/selenium-webdriver/edge.js
@@ -114,7 +114,7 @@ class Options {
     /** @private {!Object} */
     this.options_ = {};
 
-    /** @private {?capabilities.ProxyConfig} */
+    /** @private {?./lib/proxy.Config} */
     this.proxy_ = null;
   }
 
@@ -144,7 +144,7 @@ class Options {
 
   /**
    * Sets the proxy settings for the new session.
-   * @param {capabilities.ProxyConfig} proxy The proxy configuration to use.
+   * @param {./lib/proxy.Config} proxy The proxy configuration to use.
    * @return {!Options} A self reference.
    */
   setProxy(proxy) {